Add files via upload
Research paper on methods of handling missing data in research studies.

When faced with missingness in a data set it is important to carefully consider what method
to use to handle the missing values. Missing data can be classified into three main categories
as explained by Rubin (1976), MCAR, MAR and MNAR. It is essential to correctly categorise
missing data. This is because correctly identifying the type of missingness will lead to the
correct method of handling missing data. If missing values are incorrectly classified, a wrong
choice of a method for handling the missing data can be used resulting in misrepresentation
of the data. From this misleading results can be obtained for the study.